+++ packages/playground/nested-deps/vite.config.js
@@ -3,6 +3,11 @@
  */
 module.exports = {
   optimizeDeps: {
-    include: ['test-package-a', 'test-package-b']
+    include: [
+      'test-package-a',
+      'test-package-b',
+      'test-package-c',
+      'test-package-c/side'
+    ]
   }
 }