+++ packages/playground/preload/src/components/Home.vue
@@ -1,0 +1,9 @@
+<template>
+  <div>
+    This is <b>home</b> page.
+    <br />
+    Go to <router-link to="/about">About</router-link> page
+  </div>
+</template>
+
+<script setup></script>