+++ packages/vite/src/node/optimizer/index.ts
@@ -128,7 +128,7 @@ export async function optimizeDeps(
   }
 
   if (!force) {
-    let prevData
+    let prevData: DepOptimizationMetadata | undefined
     try {
       prevData = JSON.parse(fs.readFileSync(dataPath, 'utf-8'))
     } catch (e) {}
@@ -144,6 +144,12 @@ export async function optimizeDeps(
   } else {
     fs.mkdirSync(cacheDir, { recursive: true })
   }
+  // a hint for Node.js
+  // all files in the cache directory should be recognized as ES modules
+  writeFile(
+    path.resolve(cacheDir, 'package.json'),
+    JSON.stringify({ type: 'module' })
+  )
 
   let deps: Record<string, string>, missing: Record<string, string>
   if (!newDeps) {
@@ -261,6 +267,7 @@ export async function optimizeDeps(
     config.optimizeDeps?.esbuildOptions ?? {}
 
   const result = await build({
+    absWorkingDir: process.cwd(),
     entryPoints: Object.keys(flatIdDeps),
     bundle: true,
     format: 'esm',
@@ -354,12 +361,7 @@ function isSingleDefaultExport(exports: string[]) {
 
 const lockfileFormats = ['package-lock.json', 'yarn.lock', 'pnpm-lock.yaml']
 
-let cachedHash: string | undefined
-
 function getDepHash(root: string, config: ResolvedConfig): string {
-  if (cachedHash) {
-    return cachedHash
-  }
   let content = lookupFile(root, lockfileFormats) || ''
   // also take config into account
   // only a subset of config options that can affect dep optimization